Adds support for using IP addresses to specify peers
This commit adds support for using IP addresses to specify peers. It
also refactors out some networking function so that they can be used by
other files.
