+++ rmw_fastrtps_shared_cpp/CMakeLists.txt
@@ -96,6 +96,7 @@ add_library(rmw_fastrtps_shared_cpp
   src/time_utils.cpp
   src/TypeSupport_impl.cpp
   src/utils.cpp
+  src/utils/net_utils.cpp
 )
 target_include_directories(rmw_fastrtps_shared_cpp
   PUBLIC
